// src/entities/appointments/types/types.ts
export type AppointmentStatus = "pending" | "confirmed" | "cancelled" | "completed"